+++ CHANGELOG.md
@@ -7,8 +7,9 @@
 
 ### v1.0.5
 > Updated: 11.02.2026
-1. Adds automated tests around assessment report
+1. Adds automated tests around assessment report and questions navigation
 2. Package updates
+3. Signpost title changes in the report
 
 ### v1.0.4
 > Updated: 09.02.2026